metadata/requirements.txt
boto3~=1.35.66
pillow~=11.0.0
botocore~=1.35.66